Converting YOLO class ids to 0 base ids
This script convert YOLO format class id / category ids to
new class ids starts from 0 onwards. This script is required because YOLO training throws error if class ids are not aligned as 0,1,2,3 In case of using only a limited number of classes instead of using all classes in datasets this conversion can solve the issue.